+++ src/games/casino/coinflip.js
@@ -15,6 +15,7 @@ const User = require('../../models/User');
 const Guild = require('../../models/Guild');
 const { placeWager } = require('../../utils/placeWager');
 const { confirmBet } = require('../../utils/confirmBet');
+const { casinoRefusal } = require('./betGuard');
 const { delay } = require('../../utils/delay');
 const COLORS = require('../../utils/embedColors');
 const { newHandId, payHand, payoutNote, settledBalance } = require('./payout');
@@ -77,10 +78,10 @@ module.exports = {
         const side          = interaction.options.getString('side');
         const guildSettings = await Guild.findOne({ guildId: interaction.guild.id });
 
-        const casinoMaxBet = guildSettings?.economy?.casinoMaxBet ?? 0;
-        if (casinoMaxBet > 0 && bet > casinoMaxBet) {
+        const refusal = casinoRefusal(guildSettings, bet);
+        if (refusal) {
             releaseLock?.();
-            return interaction.reply({ content: `❌ The casino bet limit on this server is **${casinoMaxBet.toLocaleString()}** coins.`, flags: MessageFlags.Ephemeral });
+            return interaction.reply({ content: refusal, flags: MessageFlags.Ephemeral });
         }
 
         const user   = await User.findOne({ userId: interaction.user.id, guildId: interaction.guild.id });